# Sort only upper tail waits