Fixes #20.

Benchmarks:

```julia
julia> using PSIS

julia> x1 = randn(1_000);

julia> x2 = randn(100, 1_000, 4);
```

Before this PR:
```julia
julia> @btime $psis($x);
  89.436 μs (7 allocations: 16.75 KiB)

julia> @btime $psis($x2);
  16.521 ms (1144 allocations: 6.24 MiB)
```

After this PR:
```julia
julia> @btime $psis($x);
  71.606 μs (10 allocations: 16.86 KiB)

julia> @btime $psis($x2);
  7.827 ms (1444 allocations: 6.25 MiB)
```